backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.3
python-dateutil>=2.8.2
httpx>=0.27.0

# Database (SQLAlchemy with async support)
sqlalchemy[asyncio]>=2.0.25
aiosqlite>=0.19.0
python-dotenv>=1.0.0

# Azure AI Agents (Azure AI Foundry)
azure-ai-projects>=1.0.0b7
azure-identity>=1.15.0
openai>=1.14.0

# Azure SQL async support
aioodbc>=0.5.0
